Add common.reflection and concurrent.util to the instrumentor agent.
