+++ quasar-core/src/main/java/co/paralleluniverse/common/reflection/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package co.paralleluniverse.common.reflection;
+
+import org.osgi.annotation.bundle.Export;
